Clean up ofsetup files.
